fix(java): P2-4 align RoleChangeLog FK Integer→Long + Flyway V13
V2__oidc_user_extras.sql declared role_change_logs.target_user_id /
actor_user_id as INTEGER, but users.id is BIGINT and every other FK
in the codebase uses Long. Cosmetic alignment:

- RoleChangeLogEntity: targetUserId / actorUserId Integer → Long
- RoleChangeLogRepository: findByTargetUserIdOrderByChangedAtDesc(Long)
- UsersController: drop .intValue() casts, RoleChangeLogDto uses Long
- V13 migration: ALTER COLUMN target_user_id / actor_user_id TYPE
  BIGINT (no data loss, INTEGER values fit BIGINT)

Prod note (CLAUDE.md): Flyway is disabled in prod, so V13 needs the
manual ALTER + flyway_schema_history bump step described in the
migration file's header comment.

+++ java-backend/src/main/java/com/aiops/api/api/admin/UsersController.java
@@ -107,8 +107,8 @@ public class UsersController {
 		user = userRepository.save(user);
 
 		var logEntry = new RoleChangeLogEntity();
-		logEntry.setTargetUserId(user.getId().intValue());
-		logEntry.setActorUserId(actor != null ? Long.valueOf(actor.userId()).intValue() : null);
+		logEntry.setTargetUserId(user.getId());
+		logEntry.setActorUserId(actor != null ? Long.valueOf(actor.userId()) : null);
 		logEntry.setOldRoles(oldRolesJson);
 		logEntry.setNewRoles(user.getRoles());
 		logEntry.setReason(req.reason());
@@ -138,7 +138,7 @@ public class UsersController {
 
 	@GetMapping("/{id}/role-history")
 	public ApiResponse<List<RoleChangeLogDto>> roleHistory(@PathVariable Long id) {
-		var logs = logRepository.findByTargetUserIdOrderByChangedAtDesc(id.intValue());
+		var logs = logRepository.findByTargetUserIdOrderByChangedAtDesc(id);
 		return ApiResponse.ok(logs.stream().map(this::toLogDto).toList());
 	}
 
@@ -179,7 +179,7 @@ public class UsersController {
 
 	public record SetActiveRequest(Boolean isActive) {}
 
-	public record RoleChangeLogDto(Long id, Integer targetUserId, Integer actorUserId,
+	public record RoleChangeLogDto(Long id, Long targetUserId, Long actorUserId,
 	                                String oldRoles, String newRoles, String reason,
 	                                OffsetDateTime changedAt) {}
