+++ java-backend/src/main/java/com/aiops/api/domain/user/RoleChangeLogRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Repository
 public interface RoleChangeLogRepository extends JpaRepository<RoleChangeLogEntity, Long> {
 	List<RoleChangeLogEntity> findTop50ByOrderByChangedAtDesc();
-	List<RoleChangeLogEntity> findByTargetUserIdOrderByChangedAtDesc(Integer targetUserId);
+	List<RoleChangeLogEntity> findByTargetUserIdOrderByChangedAtDesc(Long targetUserId);
 }
